Moved strategy classes to external files in strategies folder with their supporting classes, and updated references in tests
