+++ tests/api/test_impressions_api.py
@@ -3,7 +3,8 @@
 import pytest
 from splitio.api import impressions, client, APIException
 from splitio.models.impressions import Impression
-from splitio.engine.impressions import Counter, ImpressionsMode
+from splitio.engine.impressions import ImpressionsMode
+from splitio.engine.strategies import Counter
 from splitio.client.util import get_metadata
 from splitio.client.config import DEFAULT_CONFIG
 from splitio.version import __version__
